housekeeping: move test, minunit
